Add files via upload
In this project we had the dataset comprised of Attributes which can be used to predict the severity of the accident in the city of Seattle. The output of this project can be used to predict the Severity of the accident in the locality of Seattle and also further can be used to make decisions which may lead to decrease in the number of fatalities in Seattle.